+++ src/main/java/devkor/com/teamcback/domain/user/validator/AppleValidator.java
@@ -1,44 +1,48 @@
-package devkor.com.teamcback.domain.user.service.validator;
+package devkor.com.teamcback.domain.user.validator;
 
 import static devkor.com.teamcback.global.response.ResultCode.UNAUTHORIZED;
 
+import devkor.com.teamcback.domain.user.validator.client.AppleClient;
 import devkor.com.teamcback.global.exception.GlobalException;
 import devkor.com.teamcback.global.jwt.OIDC.OIDCUtil;
 import devkor.com.teamcback.global.jwt.OIDC.dto.OIDCDecodePayload;
 import devkor.com.teamcback.global.jwt.OIDC.dto.OIDCPublicKeyDto;
 import devkor.com.teamcback.global.jwt.OIDC.dto.OIDCPublicKeysResponse;
+import io.jsonwebtoken.Header;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.cache.Cache;
-import org.springframework.cache.CacheManager;
 import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class KakaoValidator{
+public class AppleValidator {
     private final OIDCUtil oidcUtil;
-    private final KakaoClient kakaoClient;
+    private final AppleClient appleClient;
+    private static final String KID = "kid";
+    private static final String ALG = "alg";
 
-    @Value("${jwt.social.kakao.iss}")
+    @Value("${jwt.social.apple.iss}")
     private String ISS;
-    @Value("${jwt.social.kakao.aud}")
+    @Value("${jwt.social.apple.aud}")
     private String AUD;
 
     public OIDCPublicKeysResponse getCachedData() {
-        return kakaoClient.getPublicKeys();
+        return appleClient.getPublicKeys();
     }
 
     public String validateToken(String token) {
         try {
-            // 카카오 id_token 정보
-            String kid = oidcUtil.getKidFromUnsignedTokenHeader(token, AUD, ISS);
+            // id_token 정보
+            Header tokenInfo = oidcUtil.getUnsignedTokenClaims(token, AUD, ISS).getHeader();
+            String kid = (String) tokenInfo.get(KID);
+            String alg = (String) tokenInfo.get(ALG);
 
             // 공개키 가져오기
             OIDCPublicKeysResponse publicKeysResponse = getCachedData();
 
             OIDCPublicKeyDto oidcPublicKeyDto =
                 publicKeysResponse.getKeys().stream()
-                    .filter(o -> o.getKid().equals(kid))
+                    .filter(o -> o.getKid().equals(kid) && o.getAlg().equals(alg))
                     .findFirst()
                     .orElseThrow();
