+++ src/main/java/devkor/com/teamcback/domain/user/validator/client/KakaoClient.java
@@ -1,13 +1,15 @@
-package devkor.com.teamcback.domain.user.service.validator;
+package devkor.com.teamcback.domain.user.validator.client;
 
 import devkor.com.teamcback.global.jwt.OIDC.dto.OIDCPublicKeysResponse;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 
-@FeignClient(name = "appleClient", url = "https://appleid.apple.com")
-public interface AppleClient {
-    @Cacheable(value = "apple")
-    @GetMapping("/auth/keys")
+@FeignClient(name = "kakaoClient", url = "https://kauth.kakao.com")
+public interface KakaoClient {
+
+    @Cacheable(value = "kakao", key = "'data'")
+    @GetMapping("/.well-known/jwks.json")
     OIDCPublicKeysResponse getPublicKeys();
 }
+
